refactor: split openapi by provider
