+++ gpt2giga/api/gemini/content.py
@@ -6,12 +6,12 @@ from fastapi import APIRouter, Request
 from fastapi.responses import StreamingResponse
 
 from gpt2giga.app_state import get_gigachat_client
-from gpt2giga.core.logging.setup import rquid_context
-from gpt2giga.openapi_specs.gemini import (
+from gpt2giga.api.gemini.openapi import (
     gemini_batch_embed_contents_openapi_extra,
     gemini_count_tokens_openapi_extra,
     gemini_generate_content_openapi_extra,
 )
+from gpt2giga.core.logging.setup import rquid_context
 from gpt2giga.protocol.gemini import (
     GeminiAPIError,
     build_batch_embed_contents_response,
