+++ gpt2giga/api/openai/batches.py
@@ -5,15 +5,15 @@ from typing import Optional
 from fastapi import APIRouter, HTTPException, Query, Request
 
 from gpt2giga.app_state import get_batch_store, get_file_store, get_gigachat_client
+from gpt2giga.api.openai.helpers import _paginate_items
+from gpt2giga.api.openai.openapi import batches_openapi_extra
 from gpt2giga.common.exceptions import exceptions_handler
 from gpt2giga.common.request_json import read_request_json
-from gpt2giga.openapi_specs.openai import batches_openapi_extra
 from gpt2giga.protocol.batches import (
     build_openai_batch_object,
     get_batch_target,
     transform_batch_input_file,
 )
-from gpt2giga.api.openai.helpers import _paginate_items
 
 router = APIRouter(tags=["OpenAI"])
