--- gpt2giga/openapi_specs/__init__.py
@@ -1,1 +1,0 @@
-"""OpenAPI schema helper modules."""
